DRY/consolidation/reduce work on providers
